pyproject.toml
[project]
name = "mcr-masters-game"
version = "0.1.0"
description = ""
authors = [
    {name = "Hot6Mania",email = "than7683@gmail.com"}
]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.1"
python-dotenv = "^1.0.1"
pytest-dotenv = "^0.5.2"
fastapi = "^0.115.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
mypy = "^1.15.0"
ruff = "^0.9.6"
pre-commit = "^4.1.0"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.3"

[tool.poetry.scripts]
start = "scripts.cli:start_dev_server"
start-prod = "scripts.cli:start_prod_server"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "app.api.*"
disable_error_code = ["no-untyped-def"]

[tool.ruff]
line-length = 88
target-version = "py312"
fix = true
output-format = "grouped"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "C",   # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
    "COM", # trailing commas
    "SIM", # code simplification
    "ARG",  # unused arguments
    "PL",   # pylint
    "TRY",  # try-except style
]

ignore = [
    "TRY003",   # should be removed
    "PLR2004",  # Magic value using in comparision
]

[tool.ruff.lint.per-file-ignores]
"**/tests/*.py" = ["ARG001", "PLR0913", "ARG002", "SIM105", "F841", "ARG003", "ARG005"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.pytest.ini_options]
# env_files = [".env", ".env.test"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
